Cleaned Up JoinEvent.java
What I changed and why:

- Combined JoinEvent.java and JoinMessage.java. There is no reason to have the same event in 2 different classes.
- Removed the unnecessary reference to the Main class.
- Fixed indentation and white space.
- Changed bracket style. If you perfer the old way, feel free to change it back.
